# Label matcher optimizations