These are some low-hanging optimizations to regex label matchers that already exist for our line-filter matchers:
- Replace `.*` matchers with a noop
- Replace `.+` with `len(value) > 0
- Rewrite greedy regexes to be non-greedy
- replace `literals` with `strings.Contain`

Benchmarks:
```
Replace .* with noop
name            old time/op    new time/op    delta
EmptyFilters-8    45.2ns ± 3%     1.8ns ± 3%  -95.93%  (p=0.000 n=8+8)

name            old alloc/op   new alloc/op   delta
EmptyFilters-8     0.00B          0.00B          ~     (all equal)

name            old allocs/op  new allocs/op  delta
EmptyFilters-8      0.00           0.00          ~     (all equal)
```

```
Replace .+ with len > 0
name             old time/op    new time/op    delta
ExistsFilters-8    10.6ns ±13%     4.5ns ± 8%  -57.78%  (p=0.000 n=10+10)

name             old alloc/op   new alloc/op   delta
ExistsFilters-8     0.00B          0.00B          ~     (all equal)

name             old allocs/op  new allocs/op  delta
ExistsFilters-8      0.00           0.00          ~     (all equal)
```

```
Replace literal with strings.Contain
name              old time/op    new time/op    delta
LiteralFilters-8    10.6ns ± 8%     6.0ns ± 1%  -43.50%  (p=0.000 n=10+8)

name              old alloc/op   new alloc/op   delta
LiteralFilters-8     0.00B          0.00B          ~     (all equal)

name              old allocs/op  new allocs/op  delta
LiteralFilters-8      0.00           0.00          ~     (all equal)
```

greedy -> non-greedy optimizations are benchmarked in this PR: https://github.com/grafana/loki/pull/4775