@MasslessParticle 2023-03-02
This draft is superseded by: https://github.com/grafana/loki/pull/8659 which reuses our optimized line filters as label filters.